Move EagerResolution to rustc_infer::infer::resolve
